The useScaffoldContractWrite should have the implementation of the Multi version
